Move pkg/kubect/explain to staging
