+++ pkg/kubectl/BUILD
@@ -80,7 +80,6 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//pkg/kubectl/cmd:all-srcs",
-        "//pkg/kubectl/explain:all-srcs",
     ],
     tags = ["automanaged"],
 )
