+++ staging/src/k8s.io/kubectl/pkg/explain/test-swagger.json
