+++ staging/src/k8s.io/kubectl/pkg/explain/typename.go
